refactor(mcp): split tool operations into focused modules
Reduce cognitive load in the builtin tool server by separating persistence, mutations, verify, and list handlers while keeping the public operations API stable.
