--- src-tauri/src/mcp/builtin/tool/operations.rs
@@ -1,972 +1,0 @@
-use crate::mcp::builtin::error_guidance::{
-    guided_error, missing_param_error, not_found_error, ErrorCategory, SuccessHint, ToolGroup,
-};
-use crate::mcp::builtin::service_id::BuiltinServiceId;
-use crate::mcp::builtin::utils::load_session_tool_access;
-use crate::mcp::types::{MCPResult, MCPServerConfig, OAuthConfig, TransportConfig};
-use crate::repositories::mcp_server_repository::MCPServerRepository;
-use crate::repositories::AssistantRepository;
-use crate::state::get_mcp_server_repository;
-use serde_json::{json, Value};
-
-use super::queries::{get_server_config, get_server_details};
-
-use super::ToolServer;
-
-async fn save_server_config(config: &MCPServerConfig) -> Result<String, String> {
-    let repo = get_mcp_server_repository();
-    let server_name = config
-        .name
-        .as_ref()
-        .ok_or_else(|| "Server name is required".to_string())?;
-
-    // 1. Verify the configuration before saving
-    let tools =
-        crate::services::mcp_server_service::McpServerService::verify_config(config.clone())
-            .await
-            .map_err(|e| format!("Verification failed: {}", e))?;
-    let tools_json_str = crate::mcp::utils::serialize_mcp_tools(&tools);
-
-    let config_value = serde_json::to_value(config).map_err(|e| e.to_string())?;
-
-    // Try to update first (by name lookup), create if doesn't exist
-    let id = match repo.get_by_name(server_name).await {
-        Ok(Some(existing)) => {
-            // Update by ID with new config
-            repo.update(&existing.id, None, Some(config_value))
-                .await
-                .map_err(|e| format!("Failed to update MCP server config: {}", e))?
-                .id
-        }
-        Ok(None) => {
-            repo.create(server_name, config_value)
-                .await
-                .map_err(|e| format!("Failed to create MCP server config: {}", e))?
-                .id
-        }
-        Err(e) => return Err(format!("DB query error while saving server config: {}", e)),
-    };
-
-    // Update the cached tools immediately since we just verified it
-    let _ = repo
-        .update_cached_tools(&id, tools.len() as i32, tools_json_str)
-        .await;
-
-    Ok(id)
-}
-
-async fn delete_server_config_db(id_or_name: String) -> Result<(), String> {
-    let repo = get_mcp_server_repository();
-
-    // Try ID first, then name
-    let mut server = repo.get(&id_or_name).await.map_err(|e| e.to_string())?;
-    if server.is_none() {
-        server = repo
-            .get_by_name(&id_or_name)
-            .await
-            .map_err(|e| e.to_string())?;
-    }
-
-    let server = server.ok_or_else(|| format!("MCP server '{}' not found", id_or_name))?;
-
-    repo.delete(&server.id)
-        .await
-        .map_err(|e| format!("DB Delete Error: {}", e))?;
-    Ok(())
-}
-
-/// Register a new MCP server configuration
-pub async fn register_server(_server: &ToolServer, args: Value) -> Result<MCPResult, String> {
-    let name = match args.get("name").and_then(|v| v.as_str()) {
-        Some(n) if !n.is_empty() => n.to_string(),
-        Some(_) => {
-            return Ok(guided_error(
-                ErrorCategory::InvalidInput,
-                "Server name cannot be empty",
-                ToolGroup::Tool,
-            )
-            .with_guidance(vec!["Provide a unique name for this MCP server".to_string()])
-            .to_mcp_result())
-        }
-        None => return Ok(missing_param_error("name", ToolGroup::Tool)),
-    };
-
-    if BuiltinServiceId::from_alias(&name).is_some() {
-        return Ok(guided_error(
-            ErrorCategory::InvalidInput,
-            format!(
-                "Server name '{}' is reserved for a builtin service. Choose a different name.",
-                name
-            ),
-            ToolGroup::Tool,
-        )
-        .with_guidance(vec!["Use a unique name that doesn't match a builtin service (e.g. planning, browser, workspace)".to_string()])
-        .to_mcp_result());
-    }
-
-    match get_server_config(&name).await {
-        Ok(Some(_)) => {
-            return Ok(guided_error(
-                ErrorCategory::DuplicateResource,
-                format!("Server name '{}' already exists", name),
-                ToolGroup::Tool,
-            )
-            .with_guidance(vec![
-                format!(
-                    "Use tool__updateServer(name=\"{}\", transport=...) to change the existing server configuration",
-                    name
-                ),
-                format!(
-                    "Use tool__listServers(query=\"{}\") to inspect the existing server before modifying it",
-                    name
-                ),
-                "Choose a different unique name if you want to register a separate server"
-                    .to_string(),
-            ])
-            .to_mcp_result());
-        }
-        Ok(None) => {}
-        Err(error) => {
-            return Ok(guided_error(
-                ErrorCategory::DatabaseError,
-                format!(
-                    "Failed to check whether server '{}' already exists: {}",
-                    name, error
-                ),
-                ToolGroup::Tool,
-            )
-            .with_guidance(vec![
-                "The server registry could not be queried, so registration was aborted to avoid mutating an existing server by mistake".to_string(),
-                "Retry the operation after the database/service issue is resolved".to_string(),
-            ])
-            .to_mcp_result());
-        }
-    }
-
-    let transport_val = match args.get("transport") {
-        Some(t) => t,
-        Option::None => return Ok(missing_param_error("transport", ToolGroup::Tool)),
-    };
-
-    let transport: TransportConfig = match serde_json::from_value(transport_val.clone()) {
-        Ok(config) => config,
-        Err(e) => {
-            return Ok(guided_error(
-                ErrorCategory::InvalidInput,
-                format!("Invalid transport config: {}", e),
-                ToolGroup::Tool,
-            )
-            .with_guidance(vec![
-                "Verify the 'transport' object matches the expected schema (stdio or http)"
-                    .to_string(),
-                "For stdio: { \"type\": \"stdio\", \"command\": \"...\", \"args\": [...] }"
-                    .to_string(),
-                "For http: { \"type\": \"http\", \"url\": \"...\" }".to_string(),
-            ])
-            .to_mcp_result())
-        }
-    };
-
-    // Extract optional description for metadata
-    let metadata = args
-        .get("description")
-        .and_then(|v| v.as_str())
-        .map(|desc| crate::mcp::types::ServerMetadata {
-            description: Some(desc.to_string()),
-            vendor: None,
-            version: None,
-        });
-
-    let authentication = match args.get("authentication") {
-        Some(v) => match serde_json::from_value::<OAuthConfig>(v.clone()) {
-            Ok(config) => Some(config),
-            Err(e) => {
-                return Ok(guided_error(
-                    ErrorCategory::InvalidInput,
-                    format!("Invalid authentication config: {}", e),
-                    ToolGroup::Tool,
-                )
-                .to_mcp_result())
-            }
-        },
-        None => None,
-    };
-
-    let config = MCPServerConfig {
-        name: Some(name.clone()),
-        transport,
-        authentication,
-        metadata,
-    };
-
-    let id = match save_server_config(&config).await {
-        Ok(id) => id,
-        Err(e) => {
-            return Ok(guided_error(
-                ErrorCategory::DatabaseError,
-                format!("Failed to register server: {}", e),
-                ToolGroup::Tool,
-            )
-            .with_guidance(vec![
-                "Check database connectivity".to_string(),
-                "Ensure the server name is unique".to_string(),
-            ])
-            .to_mcp_result());
-        }
-    };
-
-    // Note: Session Isolation means we cannot auto-start via global manager
-    // External servers are now created per-session through MCPServiceProxyManager
-
-    // Emit resource updated event for frontend cache revalidation
-    crate::agent::tauri_events::emit_resource_updated(
-        "mcpServer",
-        "create",
-        Some(name.to_string()),
-    );
-
-    let hint = SuccessHint::new(
-        format!(
-            "✓ Server configuration saved\n\n• Server Name: {}\n• Server ID: {}\n\nStatus: Saved",
-            name, id
-        ),
-        vec![
-            "Use tool__listServers({\"availability\":\"inventory\"}) to confirm the registered server.".to_string(),
-            "Attach this Server ID to an agent config with agent__updateAgent(id:\"<agentId>\", externalMcpServers:[...]). That updates the template for future sessions only — it cannot add tools to your currently active session.".to_string(),
-            "Confirm what this session can call with tool__listServers({\"availability\":\"session\"}).".to_string(),
-        ],
-    );
-    Ok(hint.to_mcp_result_with_data(Some(json!({ "name": name, "id": id }))))
-}
-
-/// Delete an MCP server
-pub async fn delete_server(_server: &ToolServer, args: Value) -> Result<MCPResult, String> {
-    let name = match args.get("name").and_then(|v| v.as_str()) {
-        Some(n) if !n.is_empty() => n.to_string(),
-        Some(_) => {
-            return Ok(guided_error(
-                ErrorCategory::InvalidInput,
-                "Target name cannot be empty",
-                ToolGroup::Tool,
-            )
-            .with_guidance(vec!["Specify the name of the server to delete".to_string()])
-            .to_mcp_result())
-        }
-        Option::None => return Ok(missing_param_error("name", ToolGroup::Tool)),
-    };
-
-    // Get server repository and fetch server details to get both name and id
-    let repo = get_mcp_server_repository();
-    let mut server = repo.get(&name).await.map_err(|e| e.to_string())?;
-    if server.is_none() {
-        server = repo.get_by_name(&name).await.map_err(|e| e.to_string())?;
-    }
-
-    let server = match server {
-        Some(s) => s,
-        None => return Ok(not_found_error("Server", &name, ToolGroup::Tool)),
-    };
-
-    // Get all agent configs to check for dependencies
-    let assistant_repo = crate::state::get_assistant_repository();
-    let assistants = assistant_repo
-        .list_assistants()
-        .await
-        .map_err(|e| format!("Failed to list agent configurations: {}", e))?;
-
-    let mut affected_agents = Vec::new();
-    for assistant in assistants {
-        let parsed_config: Value = serde_json::from_str(&assistant.config).unwrap_or(json!({}));
-        if let Some(mcp_server_ids) = parsed_config.get("mcpServerIds").and_then(|v| v.as_array()) {
-            let contains_server = mcp_server_ids.iter().any(|id_val| {
-                if let Some(id_str) = id_val.as_str() {
-                    id_str == server.id || id_str == server.name
-                } else {
-                    false
-                }
-            });
-            if contains_server {
-                affected_agents.push(format!("{} (ID: {})", assistant.name, assistant.id));
-            }
-        }
-    }
-
-    // Note: Session Isolation means we cannot stop via global manager
-    // Servers are managed per-session, not globally
-
-    // Delete config
-    if let Err(e) = delete_server_config_db(server.name.clone()).await {
-        return Ok(guided_error(
-            ErrorCategory::DatabaseError,
-            format!("Failed to exclude server configuration: {}", e),
-            ToolGroup::Tool,
-        )
-        .with_guidance(vec![
-            "Verify database permissions".to_string(),
-            format!(
-                "Use tool__listServers({{\"availability\":\"inventory\",\"query\":\"{}\"}}) to confirm the name exists",
-                server.name
-            ),
-        ])
-        .to_mcp_result());
-    }
-
-    // Emit resource updated event for frontend cache revalidation
-    crate::agent::tauri_events::emit_resource_updated(
-        "mcpServer",
-        "delete",
-        Some(server.name.to_string()),
-    );
-
-    let mut message = format!(
-        "Excluded server '{}' (ID: {}) from configuration. WARNING: This operation is irreversible.",
-        server.name, server.id
-    );
-
-    if !affected_agents.is_empty() {
-        message.push_str("\n\n⚠️ WARNING: The following agent configurations referenced this server and will lose access to its tools:\n");
-        for agent in &affected_agents {
-            message.push_str(&format!("- {}\n", agent));
-        }
-        message.push_str("\nPlease update these agent configurations using agent__updateAgent to remove this server reference.");
-    }
-
-    let hint = SuccessHint::new(
-        message,
-        vec![
-            "Use tool__listServers({\"availability\":\"inventory\"}) to verify remaining servers"
-                .to_string(),
-            "Use agent__updateAgent to clean up orphaned server references if needed".to_string(),
-        ],
-    );
-    Ok(hint.to_mcp_result_with_data(Some(json!({
-        "name": server.name,
-        "id": server.id,
-        "status": "deleted",
-        "affectedAgents": affected_agents
-    }))))
-}
-
-/// Update an existing MCP server configuration
-pub async fn update_server(_server: &ToolServer, args: Value) -> Result<MCPResult, String> {
-    let name = match args.get("name").and_then(|v| v.as_str()) {
-        Some(n) if !n.is_empty() => n,
-        Some(_) => {
-            return Ok(guided_error(
-                ErrorCategory::InvalidInput,
-                "Target name cannot be empty",
-                ToolGroup::Tool,
-            )
-            .with_guidance(vec!["Specify the name of the server to update".to_string()])
-            .to_mcp_result())
-        }
-        Option::None => return Ok(missing_param_error("name", ToolGroup::Tool)),
-    };
-
-    if BuiltinServiceId::from_alias(name).is_some() {
-        return Ok(guided_error(
-            ErrorCategory::InvalidInput,
-            format!(
-                "Server name '{}' is reserved for a builtin service. Choose a different name.",
-                name
-            ),
-            ToolGroup::Tool,
-        )
-        .with_guidance(vec!["Use a unique name that doesn't match a builtin service (e.g. planning, browser, workspace)".to_string()])
-        .to_mcp_result());
-    }
-
-    let transport = match args.get("transport") {
-        Some(t) => t,
-        Option::None => return Ok(missing_param_error("transport", ToolGroup::Tool)),
-    };
-
-    let transport_config: TransportConfig = match serde_json::from_value(transport.clone()) {
-        Ok(config) => config,
-        Err(e) => {
-            return Ok(guided_error(
-                ErrorCategory::InvalidInput,
-                format!("Invalid transport config: {}", e),
-                ToolGroup::Tool,
-            )
-            .with_guidance(vec![
-                "Verify the 'transport' object matches the expected schema".to_string(),
-            ])
-            .to_mcp_result())
-        }
-    };
-
-    // Check if server exists (Hallucination Firewall - Section 3.2)
-    if let Ok(Option::None) = get_server_config(name).await {
-        return Ok(not_found_error("Server", name, ToolGroup::Tool));
-    }
-
-    // Extract optional description for metadata
-    let metadata = args
-        .get("description")
-        .and_then(|v| v.as_str())
-        .map(|desc| crate::mcp::types::ServerMetadata {
-            description: Some(desc.to_string()),
-            vendor: None,
-            version: None,
-        });
-
-    let authentication = match args.get("authentication") {
-        Some(v) => match serde_json::from_value::<OAuthConfig>(v.clone()) {
-            Ok(config) => Some(config),
-            Err(e) => {
-                return Ok(guided_error(
-                    ErrorCategory::InvalidInput,
-                    format!("Invalid authentication config: {}", e),
-                    ToolGroup::Tool,
-                )
-                .to_mcp_result())
-            }
-        },
-        None => None,
-    };
-
-    let config = MCPServerConfig {
-        name: Some(name.to_string()),
-        transport: transport_config,
-        authentication,
-        metadata,
-    };
-
-    // Update config and get ID
-    let id = match save_server_config(&config).await {
-        Ok(id) => id,
-        Err(e) => {
-            return Ok(guided_error(
-                ErrorCategory::DatabaseError,
-                format!("Failed to update server configuration: {}", e),
-                ToolGroup::Tool,
-            )
-            .with_guidance(vec!["Check database connectivity".to_string()])
-            .to_mcp_result());
-        }
-    };
-
-    // Note: Session Isolation means we cannot restart via global manager
-    // Configuration updates take effect when servers are next started in a session
-
-    // Emit resource updated event for frontend cache revalidation
-    crate::agent::tauri_events::emit_resource_updated(
-        "mcpServer",
-        "update",
-        Some(name.to_string()),
-    );
-
-    let hint = SuccessHint::new(
-        format!("✓ Server configuration updated for '{}' (ID: {})", name, id),
-        vec!["Use tool__listServers({\"availability\":\"inventory\",\"query\":\"<server-name>\"}) to verify changes".to_string()],
-    );
-    Ok(hint.to_mcp_result_with_data(Some(json!({ "name": name, "id": id }))))
-}
-
-/// Verify server configuration and connectivity
-pub async fn verify_server(
-    _server: &ToolServer,
-    args: Value,
-    session_id: Option<&str>,
-) -> Result<MCPResult, String> {
-    use crate::mcp::types::TransportConfig;
-    use std::time::Instant;
-
-    let name = match args.get("name").and_then(|v| v.as_str()) {
-        Some(n) => n,
-        Option::None => return Ok(missing_param_error("name", ToolGroup::Tool)),
-    };
-
-    // Get server config
-    let (id, config) = match get_server_details(name).await? {
-        Some(details) => details,
-        Option::None => return Ok(not_found_error("Server", name, ToolGroup::Tool)),
-    };
-
-    let access = load_session_tool_access(session_id).await;
-    let (session_status, session_access_line) =
-        access.external_access_report(session_id, &id, name);
-
-    // Determine transport type
-    let (transport_type, transport_details) = match &config.transport {
-        TransportConfig::Stdio { command, args, .. } => {
-            let args_str = if args.is_empty() {
-                "(no arguments)".to_string()
-            } else {
-                args.join(" ")
-            };
-            (
-                "stdio",
-                format!("Command: {}\nArguments: {}", command, args_str),
-            )
-        }
-        TransportConfig::Http { url, .. } => ("http", format!("URL: {}", url)),
-    };
-
-    // Test connection and list tools
-    let start_time = Instant::now();
-    let verification_result = test_server_connection(&config, name).await;
-    let latency_ms = start_time.elapsed().as_millis();
-
-    // Emit resource updated event for frontend cache revalidation
-    crate::agent::tauri_events::emit_resource_updated(
-        "mcpServer",
-        "verify",
-        Some(name.to_string()),
-    );
-
-    match verification_result {
-        Ok((tool_count, tools_json)) => {
-            // Persist tool list to database (count + names/descriptions)
-            let repo = get_mcp_server_repository();
-            if let Err(e) = repo
-                .update_cached_tools(&id, tool_count as i32, tools_json)
-                .await
-            {
-                log::warn!(
-                    "Failed to cache tool list for '{}' (ID: {}): {}",
-                    name,
-                    id,
-                    e
-                );
-                // Continue - don't fail verification if cache update fails
-            }
-
-            let result_text = format!(
-                "✓ Server '{}' (ID: {}) connectivity verification successful\n\n\
-                Transport: {}\n\
-                {}\n\
-                Status: Connected and responsive\n\
-                Tools discovered: {} (cached metadata — not the same as session-callable access)\n\
-                Connection latency: {}ms\n\
-                {}\n\n\
-                Verification only confirms the server config can connect. It does not enable tools in your currently active session.",
-                name,
-                id,
-                transport_type,
-                transport_details,
-                tool_count,
-                latency_ms,
-                session_access_line
-            );
-
-            Ok(SuccessHint::new(
-                result_text,
-                vec![
-                    "Use tool__listServers({\"availability\":\"session\"}) to see tools callable in this session.".to_string(),
-                    "agent__updateAgent attaches servers to an agent template for future sessions only; start a new session (or agent__startSession) to run with that access.".to_string(),
-                ],
-            )
-            .to_mcp_result_with_data(Some(json!({
-                "name": name,
-                "id": id,
-                "toolCount": tool_count,
-                "sessionStatus": session_status,
-            }))))
-        }
-        Err(error) => {
-            let error_msg = format!("✗ Server '{}' verification failed", name);
-            let error_details = format!(
-                "Transport: {}\n\
-                {}\n\
-                Status: Failed to connect or respond\n\
-                Error: {}\n\
-                Test duration: {}ms",
-                transport_type, transport_details, error, latency_ms
-            );
-
-            let suggestions = match transport_type {
-                "stdio" => vec![
-                    "Verify the command path is correct and executable".to_string(),
-                    "Check that all required arguments are provided".to_string(),
-                    "Ensure the MCP server package is installed".to_string(),
-                    "Test the command manually in terminal".to_string(),
-                ],
-                "http" => vec![
-                    "Verify the URL is correct and accessible".to_string(),
-                    "Check that the HTTP server is running".to_string(),
-                    "Ensure network connectivity to the endpoint".to_string(),
-                    "Verify authentication headers if required".to_string(),
-                ],
-                _ => vec!["Review server configuration".to_string()],
-            };
-
-            Ok(guided_error(
-                ErrorCategory::OperationFailed,
-                format!("{}\n\n{}", error_msg, error_details),
-                ToolGroup::Tool,
-            )
-            .with_guidance(suggestions)
-            .to_mcp_result())
-        }
-    }
-}
-
-/// Test server connection by spawning/connecting and calling the remote MCP ListTools API.
-/// Returns `(tool_count, tools_json)` where `tools_json` is a JSON array of
-/// `{"name": "...", "description": "..."}` entries for caching.
-async fn test_server_connection(
-    config: &crate::mcp::types::MCPServerConfig,
-    server_name: &str,
-) -> Result<(usize, String), String> {
-    let mut cloned = config.clone();
-    if cloned.name.is_none() {
-        cloned.name = Some(server_name.to_string());
-    }
-
-    let tools =
-        crate::services::mcp_server_service::McpServerService::verify_config(cloned).await?;
-    let tools_json = crate::mcp::utils::serialize_mcp_tools(&tools);
-    Ok((tools.len(), tools_json))
-}
-
-/// Unified tool discovery across builtin and external MCP servers.
-pub async fn list_tools(args: Value, session_id: Option<&str>) -> Result<MCPResult, String> {
-    use crate::mcp::types::MCPServerConfig;
-
-    let query = args
-        .get("query")
-        .and_then(|v| v.as_str())
-        .unwrap_or("")
-        .to_lowercase();
-
-    let scope = args.get("scope").and_then(|v| v.as_str()).unwrap_or("all");
-    let availability = args
-        .get("availability")
-        .and_then(|v| v.as_str())
-        .unwrap_or("inventory");
-
-    let force_verify = args
-        .get("forceVerify")
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-
-    let limit = args
-        .get("limit")
-        .and_then(|v| v.as_u64())
-        .map(|v| v.clamp(1, 100))
-        .unwrap_or(50);
-    let limit = limit.min(usize::MAX as u64) as usize;
-
-    let offset = args.get("offset").and_then(|v| v.as_u64()).unwrap_or(0);
-    let offset = offset.min(usize::MAX as u64) as usize;
-
-    let include_internal = matches!(scope, "internal" | "all");
-    let include_external = matches!(scope, "external" | "all");
-    let session_view = availability == "session";
-    let access = if session_view {
-        load_session_tool_access(session_id).await
-    } else {
-        load_session_tool_access(None).await
-    };
-
-    struct MatchedTool {
-        source: String, // "Builtin" or "External: <server_name>"
-        name: String,
-        description: String,
-        status: String,
-        external_server: Option<(String, String)>,
-    }
-
-    let mut all_matched_tools: Vec<MatchedTool> = Vec::new();
-
-    // --- Internal (builtin) tools ---
-    if include_internal {
-        for entry in crate::mcp::builtin::service_id::BUILTIN_SERVICE_REGISTRY {
-            let all_tools = crate::mcp::server::tools::get_static_tools_for_server(entry.canonical);
-            for t in all_tools {
-                if query.is_empty()
-                    || t.name.to_lowercase().contains(&query)
-                    || t.description.to_lowercase().contains(&query)
-                {
-                    let status = if session_view {
-                        let (s, _) = access.builtin_status(entry.canonical);
-                        s.to_string()
-                    } else {
-                        "".to_string()
-                    };
-
-                    all_matched_tools.push(MatchedTool {
-                        source: "Builtin".to_string(),
-                        name: t.name.clone(),
-                        description: t.description.clone(),
-                        status,
-                        external_server: None,
-                    });
-                }
-            }
-        }
-    }
-
-    // --- External (user-registered) tools ---
-    if include_external {
-        let repo = get_mcp_server_repository();
-        let models = match repo.list().await {
-            Ok(m) => m,
-            Err(e) => {
-                return Ok(guided_error(
-                    ErrorCategory::DatabaseError,
-                    format!("Failed to query MCP server list: {}", e),
-                    ToolGroup::Tool,
-                )
-                .with_guidance(vec!["Check database connectivity".to_string()])
-                .to_mcp_result())
-            }
-        };
-
-        for model in &models {
-            let config_opt: Option<MCPServerConfig> = serde_json::from_str(&model.config).ok();
-
-            // Determine tool source: live (forceVerify) or cached
-            let tools_json_str: Option<String> = if force_verify {
-                if let Some(ref config) = config_opt {
-                    match test_server_connection(config, &model.name).await {
-                        Ok((_, json_str)) => Some(json_str),
-                        Err(e) => {
-                            log::warn!("tool::list live verify failed for '{}': {}", model.name, e);
-                            None
-                        }
-                    }
-                } else {
-                    None
-                }
-            } else {
-                model.cached_tools.clone()
-            };
-
-            let cached_tools: Vec<Value> = tools_json_str
-                .as_deref()
-                .and_then(|s| serde_json::from_str(s).ok())
-                .unwrap_or_default();
-
-            let server_matches_query =
-                query.is_empty() || model.name.to_lowercase().contains(&query);
-
-            let mut matched_in_server = false;
-            for t in &cached_tools {
-                let name = t["name"].as_str().unwrap_or("?");
-                let desc = t["description"].as_str().unwrap_or("");
-
-                let name_match = name.to_lowercase().contains(&query);
-                let desc_match = desc.to_lowercase().contains(&query);
-
-                if query.is_empty() || name_match || desc_match || server_matches_query {
-                    let status = if session_view {
-                        let (s, _) = access.external_status(&model.id, &model.name);
-                        s.to_string()
-                    } else {
-                        "".to_string()
-                    };
-
-                    all_matched_tools.push(MatchedTool {
-                        source: format!("External: {}", model.name),
-                        name: name.to_string(),
-                        description: desc.to_string(),
-                        status,
-                        external_server: Some((model.name.clone(), model.id.clone())),
-                    });
-                    matched_in_server = true;
-                }
-            }
-
-            if !matched_in_server && server_matches_query {
-                let status = if session_view {
-                    let (s, _) = access.external_status(&model.id, &model.name);
-                    s.to_string()
-                } else {
-                    "".to_string()
-                };
-
-                let desc = if tools_json_str.is_none() {
-                    "(No tools cached. Run with forceVerify=true to discover tools)"
-                } else {
-                    "(No tools provided by this server)"
-                };
-
-                all_matched_tools.push(MatchedTool {
-                    source: format!("External: {}", model.name),
-                    name: "-".to_string(),
-                    description: desc.to_string(),
-                    status,
-                    external_server: Some((model.name.clone(), model.id.clone())),
-                });
-            }
-        }
-    }
-
-    let total_results = all_matched_tools.len();
-    let total_tools = all_matched_tools
-        .iter()
-        .filter(|tool| tool.name != "-")
-        .count();
-    let total_server_rows = total_results.saturating_sub(total_tools);
-
-    if total_results == 0 {
-        let hint_text = if query.is_empty() {
-            "No tools found. Use tool__registerServer to add external MCP servers.".to_string()
-        } else {
-            format!(
-                "No tools found matching '{}'. Try a broader query, scope='all', or availability='inventory'.",
-                query
-            )
-        };
-        return Ok(SuccessHint::new(
-            hint_text,
-            vec![
-                "Use scope='all' to search both builtin and external tools".to_string(),
-                "Use availability='inventory' to browse platform/server inventory regardless of current session access".to_string(),
-                "Use tool__listServers({\"availability\":\"inventory\"}) to browse all available tools".to_string(),
-            ],
-        )
-        .to_mcp_result());
-    }
-
-    if offset >= total_results {
-        return Ok(SuccessHint::new(
-            format!(
-                "Offset {} exceeds total results ({}). Try calling again with offset: 0",
-                offset, total_results
-            ),
-            vec!["Reset offset to 0".to_string()],
-        )
-        .to_mcp_result());
-    }
-
-    let paginated_tools: Vec<_> = all_matched_tools
-        .into_iter()
-        .skip(offset)
-        .take(limit)
-        .collect();
-    let mut visible_external_ids = paginated_tools
-        .iter()
-        .filter_map(|tool| tool.external_server.clone())
-        .collect::<Vec<_>>();
-    visible_external_ids.sort();
-    visible_external_ids.dedup();
-
-    let result_summary = if total_server_rows > 0 {
-        format!(
-            "{} tools and {} matching servers without cached tools",
-            total_tools, total_server_rows
-        )
-    } else {
-        format!("{} tools", total_tools)
-    };
-
-    let header = if query.is_empty() {
-        format!(
-            "Found {} (scope: {}, availability: {}):\n\n",
-            result_summary, scope, availability
-        )
-    } else {
-        format!(
-            "Found {} matching '{}' (scope: {}, availability: {}):\n\n",
-            result_summary, query, scope, availability
-        )
-    };
-
-    let mut body =
-        String::from("| Source | Tool Name | Status | Description |\n|---|---|---|---|\n");
-    for t in &paginated_tools {
-        let desc = if t.description.len() > 80 {
-            let mut end = 77;
-            while end > 0 && !t.description.is_char_boundary(end) {
-                end -= 1;
-            }
-            format!("{}...", &t.description[..end])
-        } else {
-            t.description.clone()
-        };
-        // Escape pipes and newlines for markdown tables
-        let desc = desc.replace("|", "\\|").replace('\n', " ");
-        let name = t.name.replace("|", "\\|").replace('\n', " ");
-        let source = t.source.replace("|", "\\|").replace('\n', " ");
-        let status_str = if t.status.is_empty() {
-            "-".to_string()
-        } else {
-            t.status.replace("|", "\\|").replace('\n', " ")
-        };
-
-        body.push_str(&format!(
-            "| {} | {} | {} | {} |\n",
-            source, name, status_str, desc
-        ));
-    }
-
-    if offset.saturating_add(limit) < total_results {
-        body.push_str(&format!(
-            "\n*(Showing {} to {} of {} total results. Call this tool again with offset: {} to see more)*",
-            offset + 1,
-            offset + paginated_tools.len(),
-            total_results,
-            offset.saturating_add(limit)
-        ));
-    }
-
-    let external_action = if !session_view && !visible_external_ids.is_empty() {
-        let ids_list: Vec<String> = visible_external_ids
-            .iter()
-            .map(|(name, id)| format!("  • {} → \"{}\"", name, id))
-            .collect();
-        format!(
-            "\n\n---\n📌 External server IDs (inventory only — not auto-enabled in this session):\n\
-            Server IDs found:\n(this page only)\n{}\n\n\
-            To attach them to an agent template for future sessions, call:\n  agent__updateAgent(id: \"<agentId>\", externalMcpServers: [\"<id_1>\", \"...\"])\n\n\
-            Note: agent__updateAgent cannot add or modify tools in your currently active session. Active session tool access is fixed at session start. Use availability='session' to see what you can call right now; start a new session (or agent__startSession) to run with an updated config.\n\n\
-            Use agent__listAgents(type: \"configs\") to find your target agent ID.",
-            ids_list.join("\n")
-        )
-    } else {
-        String::new()
-    };
-
-    let mut hints = if session_view {
-        vec![
-            "Session mode shows whether the current session can actually call each tool. Use availability='inventory' to browse registered platform tools regardless of current access.".to_string(),
-        ]
-    } else {
-        vec![
-            "Inventory mode lists registered platform tools; listing or verifying a server does not make it callable here. Use availability='session' for tools permitted in the current session.".to_string(),
-        ]
-    };
-    if !force_verify && include_external {
-        hints.push(
-            "Use forceVerify=true to get a live tool list from external servers (slower)"
-                .to_string(),
-        );
-    }
-
-    let structured_results = paginated_tools
-        .iter()
-        .map(|tool| {
-            json!({
-                "source": tool.source,
-                "name": tool.name,
-                "status": tool.status,
-                "description": tool.description,
-            })
-        })
-        .collect::<Vec<_>>();
-    let external_servers = visible_external_ids
-        .iter()
-        .map(|(name, id)| json!({ "name": name, "id": id }))
-        .collect::<Vec<_>>();
-
-    Ok(
-        SuccessHint::new(format!("{}{}{}", header, body, external_action), hints)
-            .to_mcp_result_with_data(Some(json!({
-                "query": query,
-                "scope": scope,
-                "availability": availability,
-                "forceVerify": force_verify,
-                "offset": offset,
-                "limit": limit,
-                "totalResults": total_results,
-                "totalTools": total_tools,
-                "totalServerRows": total_server_rows,
-                "results": structured_results,
-                "externalServers": external_servers,
-            }))),
-    )
-}
